requirements.txt
google-cloud-firestore
uuid
fastapi
uvicorn
pydantic
google-cloud-aiplatform
streamlit
requests
supervisor
